Refactored solution to separate expressions and quantities
