+++ src/expressions/IDerived{T}.cs
@@ -1,4 +1,4 @@
-namespace Engineering.Units.Expressions
+namespace Engineering.Expressions
 {
     public interface IDerived<T> where T : IExpressible
     {
