+++ src/expressions/MultiplicationExpression{T}.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Engineering.Units.Expressions
+namespace Engineering.Expressions
 {
     public sealed class MultiplicationExpression<T> : BinaryExpression<T>
         where T : IExpressible
