+++ src/quantities/Quantity.cs
@@ -1,6 +1,6 @@
-using Engineering.Units.Expressions;
+using Engineering.Expressions;
 
-namespace Engineering.Units
+namespace Engineering.Quantities
 {
     public abstract class Quantity : IQuantity
     {
