+++ src/quantities/project.json
@@ -2,13 +2,10 @@
   "version": "1.0.0-*",
   "buildOptions": {
     "debugType": "portable",
-    "outputName": "Engineering.Units"
+    "outputName": "Engineering.Quantities"
   },
   "dependencies": {
     "expressions" : {
-      "target" : "project"
-    },
-    "quantities" : {
       "target" : "project"
     }
   },
@@ -16,7 +13,7 @@
     "netstandard1.6": {
       "dependencies": {
         "NETStandard.Library": "1.6.1"
-      }
+      }   
     }
   }
 }
