+++ src/units/IUnit.cs
@@ -1,4 +1,5 @@
-﻿using Engineering.Units.Expressions;
+﻿using Engineering.Expressions;
+using Engineering.Quantities;
 
 namespace Engineering.Units
 {
