+++ src/units/si/Ampere.cs
@@ -1,3 +1,5 @@
+using Engineering.Quantities;
+
 namespace Engineering.Units.SI
 {
     public sealed class Ampere : BaseUnit
