+++ test/demo/project.json
@@ -6,6 +6,12 @@
     "emitEntryPoint": true
   },
   "dependencies": {
+    "expressions" : {
+      "target": "project"
+    },
+    "quantities" : {
+      "target": "project"
+    },
     "units" : {
       "target": "project"
     },
