+++ test/unit/project.json
@@ -8,6 +8,12 @@
     "System.Runtime.Serialization.Primitives": "4.3.0",
     "xunit": "2.1.0",
     "dotnet-test-xunit": "1.0.0-rc2-build10015",
+    "expressions" : {
+      "target": "project"
+    },
+    "quantities" : {
+      "target": "project"
+    },
     "units" : {
       "target": "project"
     },
